refactored wjs design formats
